add prop inputPlaceholder to allow input to receive placeholder
Seems counterintuitive to want to use a placeholder with floating label, but sometimes you want to display a placeholder on the input when it's in focus. This allows you to pass a placeholder to the input itself. You can then toggle whether to show or hide the placeholder based on whether it's in focus.